move btla templates to mha_dense_wrapper.h
